% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_sigmoid.R
\name{get_sigmoid_model}
\alias{get_sigmoid_model}
\title{Get the sigmoid model (intended for easy transition between development
and release)}
\usage{
get_sigmoid_model()
}
\value{
The model
}
\description{
Get the sigmoid model (intended for easy transition between development
and release)
}
