% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TABI-package.R
\name{plot_generated_gene}
\alias{plot_generated_gene}
\title{Plots the observed aganst the modelled read counts for a gene}
\usage{
plot_generated_gene(TABi_obj, gene, covariate = colnames(TABi_obj$input.X)[2])
}
\arguments{
\item{formula}{A formula}

\item{data}{A tibble}

\item{link}{A character string}
}
\value{
A tibble
}
\description{
Plots the observed aganst the modelled read counts for a gene
}
