% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TABI-package.R
\name{plot_posterior}
\alias{plot_posterior}
\title{Plots the posterior credible interval across all genes}
\usage{
plot_posterior(TABi_obj, covariate = colnames(TABi_obj$input.X)[2], CI = 1 -
  (0.05/ncol(TABi_obj$input.y)))
}
\arguments{
\item{TABi_obj}{A TABI output object}

\item{covariate}{A character string}

\item{CI}{A real number}
}
\value{
A tibble
}
\description{
Plots the posterior credible interval across all genes
}
