% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TABI-package.R
\name{plot_generated_gene}
\alias{plot_generated_gene}
\title{Plots the observed aganst the modelled read counts for a gene}
\usage{
plot_generated_gene(
  TABi_obj,
  gene,
  covariate = colnames(TABi_obj$input.X)[2],
  CI = 1 - (0.05/ncol(TABi_obj$input.y))
)
}
\arguments{
\item{TABi_obj}{A TABI output object}

\item{gene}{A character string}

\item{covariate}{A character string}

\item{CI}{A real number}
}
\value{
A tibble
}
\description{
Plots the observed aganst the modelled read counts for a gene
}
