% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_sigmoid.R
\name{sigmoid_link}
\alias{sigmoid_link}
\title{Perform generalised linear model on RNA seq data}
\usage{
sigmoid_link(
  X,
  y,
  multiplier,
  prior,
  iter,
  warmup,
  model = stanmodels$DE_sigmoid_hierarchical,
  control = list(adapt_delta = 0.8, stepsize = 0.1, max_treedepth = 10),
  prior_only = 0,
  shards = 1
)
}
\arguments{
\item{X}{A design matrix}

\item{y}{A matrix of expression}

\item{model}{can override the default model for tests/development}

\item{house_keeping_genes}{A character string}
}
\value{
A list
}
\description{
Perform generalised linear model on RNA seq data
}
