% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_sigmoid.R
\name{sigmoid_link}
\alias{sigmoid_link}
\title{Perform generalised linear model on RNA seq data}
\usage{
sigmoid_link(X, y, prior, iter, warmup)
}
\arguments{
\item{X}{A design matrix}

\item{y}{A matrix of expression}

\item{house_keeping_genes}{A character string}
}
\value{
A list
}
\description{
Perform generalised linear model on RNA seq data
}
